storybook: move manchette's stories into /storybook
