+++ storybook/stories/ui-charts/manchette/Manchette.stories.tsx
@@ -1,9 +1,9 @@
+import { Manchette } from '@osrd-project/ui-charts';
 import '@osrd-project/ui-core/dist/theme.css';
 import '@osrd-project/ui-charts/dist/theme.css';
 import type { Meta, StoryObj } from '@storybook/react';
 
 import { SAMPLE_WAYPOINTS } from './assets/sampleData';
-import Manchette from '../components/Manchette';
 
 const meta: Meta<typeof Manchette> = {
   component: Manchette,
