+++ storybook/stories/ui-charts/manchetteWithSpaceTimeChart/Menu.tsx
