--- ui-charts/src/manchette/Manchette/stories/assets/sampleData.ts
@@ -1,32 +1,0 @@
-import { type Waypoint } from '../../types';
-
-export const SAMPLE_WAYPOINTS: Waypoint[] = [
-  {
-    id: 'South_West_station',
-    name: 'South_West_station',
-    secondaryCode: 'BV',
-    position: 0,
-    weight: 100,
-  },
-  {
-    id: 'Mid_West_station',
-    name: 'Mid_West_station',
-    secondaryCode: 'BV',
-    position: 13000000,
-    weight: 30,
-  },
-  {
-    id: 'Mid_East_station',
-    name: 'Mid_East_station',
-    secondaryCode: 'BV',
-    position: 27550000,
-    weight: 50,
-  },
-  {
-    id: 'North_East_station',
-    name: 'North_East_station',
-    secondaryCode: 'BV',
-    position: 47050000,
-    weight: 100,
-  },
-];
